Tweak SBT boot jar handling
Fixes https://github.com/alexarchambault/coursier/issues/402
