+++ plugin/src/main/scala-2.10/coursier/SbtBootJars.scala
@@ -1,0 +1,18 @@
+package coursier
+
+import java.io.File
+
+object SbtBootJars {
+  def apply(
+    scalaOrg: String,
+    scalaVersion: String,
+    jars: Seq[File]
+  ): Map[(Module, String), File] =
+    jars.collect {
+      case jar if jar.getName.endsWith(".jar") =>
+        val name = jar.getName.stripSuffix(".jar")
+        val mod = Module(scalaOrg, name)
+
+        (mod, scalaVersion) -> jar
+    }.toMap
+}
