+++ plugin/src/sbt-test/sbt-coursier/scala-sources-javadoc-jars/build.sbt
@@ -1,0 +1,46 @@
+scalaVersion := appConfiguration.value.provider.scalaProvider.version
+
+coursierCachePolicies := {
+  if (sys.props("os.name").startsWith("Windows"))
+    coursierCachePolicies.value
+  else
+    Seq(coursier.CachePolicy.ForceDownload)
+}
+
+lazy val updateClassifiersCheck = TaskKey[Unit]("updateClassifiersCheck")
+
+updateClassifiersCheck := {
+
+  val configReport = updateClassifiers.value
+    .configuration("compile")
+    .getOrElse {
+      throw new Exception(
+        "compile configuration not found in updateClassifiers report"
+      )
+    }
+
+  val scalaLibraryArtifacts = configReport
+    .modules
+    .collectFirst {
+      case moduleReport
+        if moduleReport.module.organization == "org.scala-lang" &&
+             moduleReport.module.name == "scala-library" =>
+        moduleReport.artifacts
+    }
+    .toSeq
+    .flatten
+
+  def classifierArtifact(classifier: String) =
+    scalaLibraryArtifacts.collectFirst {
+      case (a, _) if a.classifier == Some(classifier) => a
+    }
+
+  def ensureHasClassifierArtifact(classifier: String) =
+    assert(
+      classifierArtifact(classifier).nonEmpty,
+      s"scala-library $classifier not found"
+    )
+
+  ensureHasClassifierArtifact("javadoc")
+  ensureHasClassifierArtifact("sources")
+}
