+++ plugin/src/sbt-test/sbt-coursier/scala-sources-javadoc-jars/src/main/scala/Main.scala
@@ -1,0 +1,6 @@
+import java.io.File
+import java.nio.file.Files
+
+object Main extends App {
+  Files.write(new File("output").toPath, "OK".getBytes("UTF-8"))
+}
