+++ plugin/src/sbt-test/sbt-coursier/scala-sources-javadoc-jars/test
@@ -1,0 +1,4 @@
+$ delete output
+> run
+$ exists output
+> updateClassifiersCheck
